Move shell functions out of surface.cpp and into shell.cpp
